refactor: Update HeroSection.astro component
This commit refactors the HeroSection.astro component by optimizing the import statements for the react-icons library. It replaces the individual imports for FiDownload and IoMailOutline with a single import statement that includes both icons from the Io5 package. Additionally, it updates the text color of the software engineer role to a lighter shade of neutral gray (text-neutral-300) for better readability. These changes improve the code organization and visual consistency of the HeroSection component.